Move SafeguardingReviewComponent under candidate namespace
