+++ app/components/candidate_interface/safeguarding_review_component.html.erb
