+++ app/views/candidate_interface/safeguarding/show.html.erb
@@ -5,7 +5,7 @@
   <%= t('page_titles.suitability_to_work_with_children') %>
 </h1>
 
-<%= render(SafeguardingReviewComponent.new(application_form: @current_application)) %>
+<%= render(CandidateInterface::SafeguardingReviewComponent.new(application_form: @current_application)) %>
 
 <%= render(CandidateInterface::CompleteSectionComponent.new(application_form: @application_form,
                                         path: candidate_interface_complete_safeguarding_path,
